requirements.txt
python-telegram-bot==20.6
aiohttp
asyncio
flask        # Optional: Only if you add a web UI
